requirements.txt
streamlit
langchain
langchain-openai
langchain-groq
langchain-chroma
pdfplumber
python-dotenv